Moved malt package into a subdirectory of doughboys-native for now. It's not used in any other packages atm and this creates less headaches for now
