--- malt/package.json
@@ -1,35 +1,0 @@
-{
-    "name": "doughboys-malt",
-    "main": "src/index",
-    "private": true,
-    "description": "Glue that binds together doughboys-web and doughboys-native",
-    "dependencies": {
-        "moment": "^2.22.2"
-    },
-    "scripts": {
-        "test": "jest"
-    },
-    "devDependencies": {
-        "babel-cli": "^6.26.0",
-        "babel-core": "^6.26.3",
-        "babel-jest": "^23.2.0",
-        "babel-plugin-transform-export-extensions": "^6.22.0",
-        "babel-plugin-transform-object-rest-spread": "^6.26.0",
-        "babel-preset-env": "^1.7.0",
-        "babel-preset-es2015": "^6.24.1",
-        "babel-preset-react": "^6.24.1",
-        "jest": "^23.3.0",
-        "webpack": "^3.12.0"
-    },
-    "jest": {
-        "transform": {
-            ".*": "<rootDir>/node_modules/babel-jest"
-        },
-        "unmockedModulePathPatterns": [
-            "<rootDir>/node_modules/react",
-            "<rootDir>/node_modules/react-dom",
-            "<rootDir>/node_modules/react-addons-test-utils",
-            "<rootDir>/node_modules/fbjs"
-        ]
-    }
-}
