+++ native/package.json
@@ -20,7 +20,6 @@
     "preset": "jest-expo"
   },
   "dependencies": {
-    "doughboys-malt": "file:../malt",
     "expo": "^27.0.1",
     "react": "16.3.1",
     "react-native": "~0.55.2"
